backend/requirements.txt
auth==0.5.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.11
dnspython==2.7.0
eventlet==0.37.0
falcon==3.1.3
Flask==3.0.3
Flask-Auth==0.85
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.29.1
mysqlclient==2.2.4
packaging==24.1
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
